netlify/functions/requirements.txt
# Requirements for the Netlify Python webhook function
# Netlify reads this file automatically for python3.11 functions
python-telegram-bot==20.1
python-dotenv==1.1.0
httpx>=0.23.3
httpcore>=0.16.3
h11>=0.14.0
anyio>=4.9.0
sniffio>=1.3.1
certifi>=2025.4.26
idna>=3.10
typing_extensions>=4.13.2
